Add wrap container for BreedPicturesPage
